# [IMP] core: improve performance of Field.__get__()