Specialize Field.__get__() (breaks the cache abstraction) in order to reduce the performance penalty of Odoo's field cache as much as possible.
Also add _data_get to the Cache class to avoid huge number of __getattribute__().

Performance improvements: Before `self.login` took 405 ns ± 0.437 ns (with the previous optimization), now it takes 292 ns ± 0.375 ns.

We think it is worth it to break the abstraction of cache in this particular case because it is one of the most used method of Odoo.