@robodoo 2024-06-10
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/168646.png)](https://mergebot.odoo.com/odoo/odoo/pull/168646)